Rewrite cli entry point to support reporting errors with exceptions
We have two types of errors:
1) "user errors" where the requested action cannot be performed and the user needs to receive the explaining message on why this is the case
2) "programming error" where we have unintended behaviour which should be fixed

The Pythonic way is to raise an Exception and let it propagate all the way up the stack. We then catch them within the entry point (main.py:main) and show the error message to the user.

The two types of Exceptions will correspond to the two types of errors:
1) DataCatsError class instances and its children for "user errors"
2) All the other exceptions will their specific types for "programming errors"

Returning 1 and 0 codes for errors is not very friendly and we probably should move away from that.