Assign descriptive syslog-tags for new containers with syslog log-driver
Docker supports log-drivers:
a config option for containers that enables redirecting that
container's logs somewhere else (as opposed to just dumping them in a json file
within the container's directory).

We have a flag to support syslog log-driver for new containers.
However, by default the program field would only include container ID
which is not very helpful when browsing the logs.

Docker 1.8 and higher is to have support to use custom syslog tags,
by redefining docker log options (--log-opt). This thing uses that
and assigns container name for all cases when syslog log-driver is turned on.