Add Progress Bar renderer for CLI
CliProgressTracker helps render a TTY progress bar
and show progress to the user.

The update_state function signature is similar to the one
that is used by celery so the advantage is that
functions that use this progress progress bar
can be tied up to celery easily.

Here is an example of use:
    with CliProgressTracker(
        task_title="", total=13) as pt:
        for i in range(14):
            pt.update_state(
                state='PROGRESS',
                meta={
                    'current': i,
                    'total': 13,
                    'status':'sup'}
                )